+++ src/Dilcore.MongoDB/Extensions/ServiceCollectionExtensions.cs
@@ -4,6 +4,7 @@ using Dilcore.MongoDB.Abstractions.Json;
 using Dilcore.MongoDB.Abstractions.Keys;
 using Dilcore.MongoDB.Abstractions.Namespace;
 using Dilcore.MongoDB.Abstractions.Options;
+using Dilcore.MongoDB.Abstractions.Provisioning;
 using Dilcore.MongoDB.Abstractions.Repositories;
 using Dilcore.MongoDB.Abstractions.Transactions;
 using Dilcore.MongoDB.DependencyInjection;
@@ -46,6 +47,7 @@ public static class ServiceCollectionExtensions
                 sp.GetRequiredService<MongoRegistrationGraph>()));
         services.AddScoped<IMongoDatabaseResolver, MongoDatabaseResolver>();
         services.AddScoped<IMongoDbCollectionFactory, MongoDbCollectionFactory>();
+        services.AddScoped<IMongoDbProvisioner, MongoDbProvisioner>();
         services.AddSingleton<IBsonJsonConverter, BsonJsonConverter>();
         services.AddScoped<IMongoDbTransactionRunner, MongoDbTransactionRunner>();
         services.AddScoped<IRepositoryResolver, RepositoryResolver>();